+++ src/orquestra/sdk/_base/cli/_dorq/_dumpers.py
@@ -12,15 +12,47 @@ from orquestra.sdk._base import serde
 from orquestra.sdk.schema.workflow_run import TaskInvocationId, WorkflowRunId
 
 
-class ArtifactDumper:
+class WFOutputDumper:
     """
-    Writes artifacts to files.
+    Writes a workflow run's output artifacts to a file.
     """
 
     def dump(
         self,
         value: t.Any,
         wf_run_id: WorkflowRunId,
+        output_index: int,
+        dir_path: Path,
+    ) -> serde.DumpDetails:
+        """
+        Serialize a single artifact value and save it as a new file.
+
+        Missing directories will be created. Filenames will be generate based on
+        ``wf_run_id`` and ``output_index``. The serialization format and file
+        extensions are inferred based on the object.
+
+        No standard errors are expected to be raised.
+        """
+
+        dump_dir = dir_path / wf_run_id / "wf_results"
+
+        dump_details = serde.dump_to_file(
+            value=value, dir_path=dump_dir, file_name_prefix=str(output_index)
+        )
+
+        return dump_details
+
+
+class TaskOutputDumper:
+    """
+    Writes task run's output artifact to a file.
+    """
+
+    def dump(
+        self,
+        value: t.Any,
+        wf_run_id: WorkflowRunId,
+        task_inv_id: TaskInvocationId,
         output_index: int,
         dir_path: Path,
     ) -> serde.DumpDetails:
@@ -34,10 +66,12 @@ class ArtifactDumper:
         No standard errors are expected to be raised.
         """
 
+        dump_dir = dir_path / wf_run_id / "task_results" / task_inv_id
+
         dump_details = serde.dump_to_file(
             value=value,
-            dir_path=dir_path,
-            file_name_prefix=f"{wf_run_id}_{output_index}",
+            dir_path=dump_dir,
+            file_name_prefix=str(output_index),
         )
 
         return dump_details