+++ tests/cli/dorq/task/test_task_results.py
@@ -1,31 +1,37 @@
 ################################################################################
-# © Copyright 2022-2023 Zapata Computing Inc.
+# © Copyright 2023 Zapata Computing Inc.
 ################################################################################
 """
-Unit tests for 'orq task logs' glue code.
+Unit tests for 'orq task results' glue code.
 """
 
 from pathlib import Path
 from unittest.mock import create_autospec
+
+import pytest
 
 from orquestra.sdk._base.cli._dorq._arg_resolvers import (
     TaskInvIDResolver,
     WFConfigResolver,
     WFRunResolver,
 )
-from orquestra.sdk._base.cli._dorq._dumpers import LogsDumper
+from orquestra.sdk._base.cli._dorq._dumpers import TaskOutputDumper
 from orquestra.sdk._base.cli._dorq._repos import WorkflowRunRepo
-from orquestra.sdk._base.cli._dorq._task import _logs
-from orquestra.sdk._base.cli._dorq._ui._presenters import WrappedCorqOutputPresenter
+from orquestra.sdk._base.cli._dorq._task import _results
+from orquestra.sdk._base.cli._dorq._ui._presenters import (
+    ArtifactPresenter,
+    WrappedCorqOutputPresenter,
+)
+from orquestra.sdk._base.serde import DumpDetails
 
 
 class TestAction:
     """
     Test boundaries::
-        [_output.Action]->[arg resolvers]
-                        ->[repos]
-                        ->[dumper]
-                        ->[presenter]
+        [_results.Action]->[arg resolvers]
+                         ->[repos]
+                         ->[dumper]
+                         ->[presenter]
     """
 
     class TestDataPassing:
@@ -44,29 +50,31 @@ class TestAction:
             fn_name = "<my task fn name>"
 
             # Resolved values
-            resolved_id = "<resolved ID>"
+            resolved_wf_run_id = "<resolved ID>"
             resolved_config = "<resolved config>"
-            resolved_invocation_id = "<resolved invocation id>"
+            resolved_inv_id = "<resolved inv id>"
 
             # Mocks
-            presenter = create_autospec(WrappedCorqOutputPresenter)
-            dumper = create_autospec(LogsDumper)
+            error_presenter = create_autospec(WrappedCorqOutputPresenter)
+            artifact_presenter = create_autospec(ArtifactPresenter)
+            dumper = create_autospec(TaskOutputDumper)
             wf_run_repo = create_autospec(WorkflowRunRepo)
 
-            fake_logs = {"task_inv": ["my_log_1", "my_log_2"]}
-            wf_run_repo.get_task_logs.return_value = fake_logs
+            fake_outputs = ["my_log_1", "my_log_2"]
+            wf_run_repo.get_task_outputs.return_value = fake_outputs
 
             config_resolver = create_autospec(WFConfigResolver)
             config_resolver.resolve.return_value = resolved_config
 
             wf_run_resolver = create_autospec(WFRunResolver)
-            wf_run_resolver.resolve_id.return_value = resolved_id
+            wf_run_resolver.resolve_id.return_value = resolved_wf_run_id
 
             task_inv_id_resolver = create_autospec(TaskInvIDResolver)
-            task_inv_id_resolver.resolve.return_value = resolved_invocation_id
+            task_inv_id_resolver.resolve.return_value = resolved_inv_id
 
-            action = _logs.Action(
-                presenter=presenter,
+            action = _results.Action(
+                error_presenter=error_presenter,
+                artifact_presenter=artifact_presenter,
                 dumper=dumper,
                 wf_run_repo=wf_run_repo,
                 config_resolver=config_resolver,
@@ -84,27 +92,31 @@ class TestAction:
             )
 
             # Then
-            # We should pass input CLI args to config resolver.
-            presenter.show_error.assert_not_called()
+            assert error_presenter.method_calls == []
 
+            # We should pass input CLI args to config resolver.
             config_resolver.resolve.assert_called_with(wf_run_id, config)
 
             # We should pass resolved_config to run ID resolver.
             wf_run_resolver.resolve_id.assert_called_with(wf_run_id, resolved_config)
 
             task_inv_id_resolver.resolve.assert_called_with(
-                task_inv_id, fn_name, resolved_id, resolved_config
+                task_inv_id, fn_name, resolved_wf_run_id, resolved_config
             )
 
             # We should pass resolved values to run repo.
-            wf_run_repo.get_task_logs.assert_called_with(
-                wf_run_id=resolved_id,
+            wf_run_repo.get_task_outputs.assert_called_with(
+                wf_run_id=resolved_wf_run_id,
+                task_inv_id=resolved_inv_id,
                 config_name=resolved_config,
-                task_inv_id=resolved_invocation_id,
             )
 
-            # We expect printing the workflow run returned from the repo.
-            presenter.show_logs.assert_called_with(fake_logs)
+            # We expect printing the outputs returned from the repo.
+            artifact_presenter.show_task_outputs.assert_called_with(
+                values=fake_outputs,
+                wf_run_id=resolved_wf_run_id,
+                task_inv_id=resolved_inv_id,
+            )
 
             # We don't expect any dumps.
             assert dumper.mock_calls == []
@@ -120,33 +132,37 @@ class TestAction:
             fn_name = "<my task fn name>"
 
             # Resolved values
-            resolved_id = "<resolved ID>"
+            resolved_wf_run_id = "<resolved ID>"
             resolved_config = "<resolved config>"
-            resolved_invocation_id = "<resolved invocation id>"
+            resolved_inv_id = "<resolved inv id>"
 
             # Mocks
-            presenter = create_autospec(WrappedCorqOutputPresenter)
+            error_presenter = create_autospec(WrappedCorqOutputPresenter)
+            artifact_presenter = create_autospec(ArtifactPresenter)
 
-            path_to_logs = "returns whatever"
-            dumper = create_autospec(LogsDumper)
-            dumper.dump.return_value = path_to_logs
+            # path_to_logs = "returns whatever"
+            # details = create_autospec(DumpDetails)
+            dump_details = "<dump details sentinel>"
+            dumper = create_autospec(TaskOutputDumper)
+            dumper.dump.return_value = dump_details
 
             wf_run_repo = create_autospec(WorkflowRunRepo)
 
-            fake_logs = {"task_inv": ["my_log_1", "my_log_2"]}
-            wf_run_repo.get_task_logs.return_value = fake_logs
+            fake_outputs = ["my_log_1", "my_log_2"]
+            wf_run_repo.get_task_outputs.return_value = fake_outputs
 
             config_resolver = create_autospec(WFConfigResolver)
             config_resolver.resolve.return_value = resolved_config
 
             wf_run_resolver = create_autospec(WFRunResolver)
-            wf_run_resolver.resolve_id.return_value = resolved_id
+            wf_run_resolver.resolve_id.return_value = resolved_wf_run_id
 
             task_inv_id_resolver = create_autospec(TaskInvIDResolver)
-            task_inv_id_resolver.resolve.return_value = resolved_invocation_id
+            task_inv_id_resolver.resolve.return_value = resolved_inv_id
 
-            action = _logs.Action(
-                presenter=presenter,
+            action = _results.Action(
+                error_presenter=error_presenter,
+                artifact_presenter=artifact_presenter,
                 dumper=dumper,
                 wf_run_repo=wf_run_repo,
                 config_resolver=config_resolver,
@@ -164,25 +180,46 @@ class TestAction:
             )
 
             # Then
+            assert error_presenter.method_calls == []
+
             # We should pass input CLI args to config resolver.
-            presenter.show_error.assert_not_called()
             config_resolver.resolve.assert_called_with(wf_run_id, config)
 
             # We should pass resolved_config to run ID resolver.
             wf_run_resolver.resolve_id.assert_called_with(wf_run_id, resolved_config)
 
             task_inv_id_resolver.resolve.assert_called_with(
-                task_inv_id, fn_name, resolved_id, resolved_config
+                task_inv_id, fn_name, resolved_wf_run_id, resolved_config
             )
 
             # We should pass resolved values to run repo.
-            wf_run_repo.get_task_logs.assert_called_with(
-                wf_run_id=resolved_id,
+            wf_run_repo.get_task_outputs.assert_called_with(
+                wf_run_id=resolved_wf_run_id,
+                task_inv_id=resolved_inv_id,
                 config_name=resolved_config,
-                task_inv_id=resolved_invocation_id,
             )
 
-            # Do not print logs to stdout
-            presenter.show_logs.assert_not_called()
-            presenter.show_dumped_wf_logs.assert_called_with(path_to_logs)
-            dumper.dump.assert_called_with(fake_logs, resolved_id, download_dir)
+            # Do not print artifact summary.
+            artifact_presenter.show_task_outputs.assert_not_called()
+
+            # Dump the two artifacts.
+            dumper.dump.assert_any_call(
+                value=fake_outputs[0],
+                wf_run_id=resolved_wf_run_id,
+                task_inv_id=resolved_inv_id,
+                output_index=0,
+                dir_path=download_dir,
+            )
+            dumper.dump.assert_any_call(
+                value=fake_outputs[1],
+                wf_run_id=resolved_wf_run_id,
+                task_inv_id=resolved_inv_id,
+                output_index=1,
+                dir_path=download_dir,
+            )
+
+            # Print two dump details, one for each task output.
+            artifact_presenter.show_dumped_artifact.assert_called_with(dump_details)
+            assert artifact_presenter.show_dumped_artifact.call_count == len(
+                fake_outputs
+            )