+++ tests/cli/dorq/test_dumpers.py
@@ -10,7 +10,7 @@ from orquestra.sdk._base.cli._dorq import _dumpers
 from orquestra.sdk.schema.ir import ArtifactFormat
 
 
-class TestArtifactDumper:
+class TestWFOutputDumper:
     @staticmethod
     @pytest.mark.parametrize(
         "value",
@@ -23,7 +23,7 @@ class TestArtifactDumper:
         # Given
         wf_run_id = "wf.1234"
         output_index = 1
-        dumper = _dumpers.ArtifactDumper()
+        dumper = _dumpers.WFOutputDumper()
 
         # When
         details = dumper.dump(
@@ -34,12 +34,22 @@ class TestArtifactDumper:
         )
 
         # Then
-        # Creates file
+        # Creates a subdir for workflow run
         children = list(tmp_path.iterdir())
         assert len(children) == 1
+        assert children[0].name == wf_run_id
+
+        # Creates a subdir with a static name
+        grandchildren = list(children[0].iterdir())
+        assert len(grandchildren) == 1
+        assert grandchildren[0].name == "wf_results"
+
+        # Creates a file for the artifact
+        grand2children = list(grandchildren[0].iterdir())
+        assert len(grand2children) == 1
 
         # Sensible dump details
-        assert details.file_path == children[0]
+        assert details.file_path == grand2children[0]
         assert details.format in {ArtifactFormat.JSON, ArtifactFormat.ENCODED_PICKLE}
 
         # Created extension
@@ -50,9 +60,9 @@ class TestArtifactDumper:
         # Given
         value = {"general": "kenobi"}
         wf_run_id = "wf.1234"
-        output_index = 1
+        output_index = 2
         dir_path = tmp_path / "new_dir"
-        dumper = _dumpers.ArtifactDumper()
+        dumper = _dumpers.WFOutputDumper()
 
         # When
         details = dumper.dump(
@@ -63,12 +73,78 @@ class TestArtifactDumper:
         )
 
         # Then
-        # Creates file
+        # Creates a subdir for workflow run
         children = list(dir_path.iterdir())
         assert len(children) == 1
+        assert children[0].name == wf_run_id
+
+        # Creates a subdir with a static name
+        grandchildren = list(children[0].iterdir())
+        assert len(grandchildren) == 1
+        assert grandchildren[0].name == "wf_results"
+        assert grandchildren[0].is_dir()
+
+        # Creates a file for the artifact
+        grand2children = list(grandchildren[0].iterdir())
+        assert len(grand2children) == 1
+        assert grand2children[0].name == "2.json"
+        assert grand2children[0].is_file()
 
         # Sensible dump details
-        assert details.file_path == children[0]
+        assert details.file_path == grand2children[0]
+        assert details.format in {ArtifactFormat.JSON, ArtifactFormat.ENCODED_PICKLE}
+
+
+class TestTaskOutputDumper:
+    @staticmethod
+    @pytest.mark.parametrize(
+        "value",
+        [
+            pytest.param({"hello": "there"}, id="jsonable_value"),
+            pytest.param(object(), id="non_jsonable_value"),
+        ],
+    )
+    def test_dir_exists(tmp_path: Path, value):
+        # Given
+        wf_run_id = "wf.1234"
+        task_inv_id = "inv3"
+        output_index = 5
+        dumper = _dumpers.TaskOutputDumper()
+
+        # When
+        details = dumper.dump(
+            value=value,
+            wf_run_id=wf_run_id,
+            task_inv_id=task_inv_id,
+            output_index=output_index,
+            dir_path=tmp_path,
+        )
+
+        # Then
+        # Creates a subdir for workflow run
+        children = list(tmp_path.iterdir())
+        assert len(children) == 1
+        assert children[0].name == wf_run_id
+
+        # Creates a subdir with a static name
+        grandchildren = list(children[0].iterdir())
+        assert len(grandchildren) == 1
+        assert grandchildren[0].name == "task_results"
+        assert grandchildren[0].is_dir()
+
+        # Creates a directory for the invocation
+        grand2children = list(grandchildren[0].iterdir())
+        assert len(grand2children) == 1
+        assert grand2children[0].name == task_inv_id
+        assert grand2children[0].is_dir()
+
+        # Creates a file for the artifact
+        grand3children = list(grand2children[0].iterdir())
+        assert len(grand3children) == 1
+        assert grand3children[0].is_file()
+
+        # Sensible dump details
+        assert details.file_path == grand3children[0]
         assert details.format in {ArtifactFormat.JSON, ArtifactFormat.ENCODED_PICKLE}
 
         # Created extension