+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * New CLI command: `python -m orquestra.sdk._base.cli._dorq._entry workflow list`. It displays a list of workflow runs corresponding to specified configurations and filters.
 * New CLI command: `python -m orquestra.sdk._base.cli._dorq._entry workflow logs`.
 * New CLI command: `python -m orquestra.sdk._base.cli._dorq._entry task logs`.
+* New CLI command: `python -m orquestra.sdk._base.cli._dorq._entry task results`.
 
 🐛 *Bug Fixes*